get rid of Qt folder
