--- src/airmap/qt/aircrafts.h
@@ -1,44 +1,0 @@
-// AirMap Platform SDK
-// Copyright © 2018 AirMap, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the License);
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//   http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an AS IS BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-#ifndef AIRMAP_QT_AIRCRAFTS_H_
-#define AIRMAP_QT_AIRCRAFTS_H_
-
-#include <airmap/aircrafts.h>
-#include <airmap/client.h>
-#include <airmap/qt/dispatcher.h>
-
-#include <memory>
-#include <string>
-
-namespace airmap {
-namespace qt {
-
-class Aircrafts : public airmap::Aircrafts, public std::enable_shared_from_this<Aircrafts> {
- public:
-  static std::shared_ptr<Aircrafts> create(const std::shared_ptr<Dispatcher>& dispatcher,
-                                           const std::shared_ptr<airmap::Client>& client);
-
-  void manufacturers(const Manufacturers::Parameters& parameters, const Manufacturers::Callback& cb) override;
-  void models(const Models::Parameters& parameters, const Models::Callback& cb) override;
-  void model_for_id(const ModelForId::Parameters& parameters, const ModelForId::Callback& cb) override;
-
- private:
-  explicit Aircrafts(const std::shared_ptr<Dispatcher>& dispatcher, const std::shared_ptr<airmap::Client>& client);
-  std::shared_ptr<Dispatcher> dispatcher_;
-  std::shared_ptr<airmap::Client> client_;
-};
-
-}  // namespace qt
-}  // namespace airmap
-
-#endif  // AIRMAP_QT_AIRCRAFTS_H_
